+++ provider/plugin_framework.go
@@ -48,6 +48,7 @@ import (
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/ml/anomalydetectionjob"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/ml/calendar"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/ml/calendar_event"
+	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/ml/calendar_job"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/ml/datafeed"
 	datafeedstate "github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/ml/datafeed_state"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/ml/filter"
@@ -247,6 +248,7 @@ func (p *Provider) resources(_ context.Context) []func() resource.Resource {
 		anomalydetectionjob.NewAnomalyDetectionJobResource,
 		calendar.NewCalendarResource,
 		calendar_event.NewCalendarEventResource,
+		calendar_job.NewCalendarJobResource,
 		filter.NewFilterResource,
 		security_detection_rule.NewSecurityDetectionRuleResource,
 		jobstate.NewMLJobStateResource,